refactor[#470] : 프로필 페이지 컴포넌트에서 토큰이 없을 경우 로그인 페이지로 이동
