+++ src/pages/[nickname]/page.tsx
@@ -1,4 +1,8 @@
+import { useEffect } from "react";
+import { useNavigate } from "react-router-dom";
+import { ROUTER_PATH } from "@/shared/constants";
 import { useNicknameParams } from "@/shared/lib";
+import { useAuthStore } from "@/shared/store";
 import { MyProfilePage } from "./myProfilePage";
 import { OtherProfilePage } from "./otherProfilePage";
 
@@ -7,7 +11,15 @@ import { OtherProfilePage } from "./otherProfilePage";
  * @:nickname: 해당 닉네임을 가진 사용자의 프로필 페이지, 만약 AuthStore에 저장된 닉네임과 같다면 마이페이지처럼 이용 가능 합니다.
  */
 export const ProfilePage = () => {
+  const token = useAuthStore((state) => state.token);
+  const navigate = useNavigate();
   const { isMyPage } = useNicknameParams();
+
+  useEffect(() => {
+    if (!token) {
+      navigate(ROUTER_PATH.LOGIN);
+    }
+  }, [token, navigate]);
 
   if (isMyPage) {
     return <MyProfilePage />;
